feat: Update nearby search function to include rank preference
The nearby search function in the places module has been updated to include a rank preference parameter. This allows the user to specify whether they want the results to be sorted by popularity or distance. The default value is set to 'POPULARITY'. This change improves the flexibility and customization of the nearby search functionality.