feat: Update PlaceTypeSelect component to include an optional id prop
The PlaceTypeSelect component in the admin area has been updated to include an optional id prop. This allows for more flexibility in assigning unique identifiers to the component instances. The id prop can be used for various purposes such as accessibility or styling. This change enhances the reusability and customization of the component.